# packages/typespec/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "www-template API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/auth/passkey/finish": {
      "post": {
        "operationId": "finishPasskeyAuthentication",
        "summary": "Finishes passkey authentication and returns a bearer session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyFinishRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/passkey/start": {
      "post": {
        "operationId": "startPasskeyAuthentication",
        "summary": "Starts a passkey authentication ceremony",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOperationErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable.",
            "headers": {
              "Cache-Control": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": ["no-store"]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailureResponse"
                }
              }
            }
          }
        },
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyStartRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthFailureClassification": {
        "type": "string",
        "enum": ["unauthenticated", "session-expired", "account-suspended", "internal-error"]
      },
      "AuthFailureResponse": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "error": {
            "$ref": "#/components/schemas/AuthFailureClassification"
          }
        }
      },
      "AuthOperationErrorResponse": {
        "type": "object",
        "required": ["requestId", "error"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "AuthSessionResponse": {
        "type": "object",
        "required": [
          "requestId",
          "accountId",
          "passkeyCredentialId",
          "sessionId",
          "accessToken",
          "expiresAt"
        ],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "accountId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "passkeyCredentialId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "sessionId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PasskeyFinishRequest": {
        "type": "object",
        "required": ["credential"],
        "properties": {
          "credential": {
            "$ref": "#/components/schemas/WebAuthnAssertionCredential"
          }
        }
      },
...
      "PasskeyStartRequest": {
        "type": "object",
        "required": ["identifier"],
        "properties": {
          "identifier": {
            "type": "string"
          }
        }
      },
      "PasskeyStartResponse": {
        "type": "object",
        "required": ["requestId", "challenge", "rpId", "userVerification"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/UlidId"
          },
          "challenge": {
            "type": "string",
            "description": "base64url-encoded challenge bytes issued by the server."
          },
          "rpId": {
            "type": "string"
          },
          "timeout": {
            "type": "integer",
            "format": "int64",
            "description": "Timeout in milliseconds suggested by the server."
          },
          "allowCredentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebAuthnCredentialDescriptor"
            },
            "description": "Allowed credential descriptors (empty for usernameless flows)."
          },
          "userVerification": {
            "type": "string",
            "enum": ["required"]
          }
        },
        "description": "WebAuthn PublicKeyCredentialRequestOptions returned by the server after BeginLogin."
      },
...
      "UlidId": {
        "type": "string",
        "description": "Canonical ULID string used for auth-owned resource and correlation identifiers."
      },
...
      "WebAuthnAssertionCredential": {
        "type": "object",
        "required": ["id", "rawId", "type", "response"],
        "properties": {
          "id": {
            "type": "string"
          },
          "rawId": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/WebAuthnAssertionResponse"
          },
          "authenticatorAttachment": {
            "type": "string"
          }
        },
        "description": "WebAuthn PublicKeyCredential for login (navigator.credentials.get result)."
      },
      "WebAuthnAssertionResponse": {
        "type": "object",
        "required": ["clientDataJSON", "authenticatorData", "signature"],
        "properties": {
          "clientDataJSON": {
            "type": "string",
            "description": "base64url-encoded clientDataJSON."
          },
          "authenticatorData": {
            "type": "string",
            "description": "base64url-encoded authenticatorData."
          },
          "signature": {
            "type": "string",
            "description": "base64url-encoded signature."
          },
          "userHandle": {
            "type": "string",
            "description": "base64url-encoded userHandle (optional)."
          }
        },
        "description": "WebAuthn AuthenticatorAssertionResponse (login)."
      },
...
      "WebAuthnCredentialDescriptor": {
        "type": "object",
        "required": ["type", "id"],
        "properties": {
          "type": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "description": "base64url-encoded credential ID."
          },
          "transports": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Credential descriptor used in allowCredentials and excludeCredentials lists."
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT 形式のアクセストークンを使用した Bearer 認証。Authorization ヘッダーに `Bearer <JWT>` を付与すること。トークンは短命の JWT であり、accountID・sessionID・iat・exp を含む。"
      }
    }
  },
  "servers": [
    {
      "url": "https://product.www-template.example",
      "description": "Product API server",
      "variables": {}
    }
  ]
}